Close #90 add pagination, move queries to repo, use comboboxes, create separate file for person routes
