+++ app/Models/MedicalEvents/Mongo/Condition.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Models\MedicalEvents\Mongo;
+
+//use MongoDB\Laravel\Eloquent\Model;
+use Illuminate\Database\Eloquent\Model;
+
+class Condition extends Model
+{
+    //    protected $connection = 'mongodb';
+    protected $guarded = [];
+}
