+++ resources/views/livewire/care-plan/parts/modals/medical-device-form-drawer.blade.php
@@ -48,9 +48,9 @@
                     <label class="label">
                         {{ __('care-plan.program') }}
                     </label>
-                    <input type="text" 
-                           class="input bg-gray-50 dark:bg-gray-700 cursor-not-allowed" 
-                           value="{{ !empty($activityForm['program']) ? ($dictionaries['medical_programs'][$activityForm['program']] ?? $activityForm['program']) : __('care-plan.medical_guarantees_program') }}" 
+                    <input type="text"
+                           class="input bg-gray-50 dark:bg-gray-700 cursor-not-allowed"
+                           value="{{ !empty($activityForm['program']) ? ($dictionaries['medical_programs'][$activityForm['program']] ?? $activityForm['program']) : __('care-plan.medical_guarantees_program') }}"
                            disabled
                     />
                 </div>
@@ -58,9 +58,9 @@
                     <label class="label">
                         {{ __('care-plan.medical_device') }}*
                     </label>
-                    <input type="text" 
-                           class="input bg-gray-50 dark:bg-gray-700 cursor-not-allowed font-medium text-gray-900 dark:text-white" 
-                           value="{{ !empty($selectedProduct) ? ($selectedProduct['name'] ?? $selectedProduct['device_names'][0]['name'] ?? $selectedProduct['description'] ?? '') : '' }}" 
+                    <input type="text"
+                           class="input bg-gray-50 dark:bg-gray-700 cursor-not-allowed font-medium text-gray-900 dark:text-white"
+                           value="{{ !empty($selectedProduct) ? ($selectedProduct['name'] ?? $selectedProduct['device_names'][0]['name'] ?? $selectedProduct['description'] ?? '') : '' }}"
                            disabled
                     />
                     <input type="hidden" wire:model="activityForm.product_reference" />
@@ -96,7 +96,6 @@
                                class="input peer ps-10 datepicker-input"
                                placeholder="02.04.2025"
                                datepicker-autohide
-                               datepicker-format="dd.mm.yyyy"
                                datepicker-button="false"
                                wire:model.live="activityForm.scheduled_period_start"
                         />
@@ -148,7 +147,6 @@
                                class="input peer ps-10 datepicker-input"
                                placeholder="02.08.2025"
                                datepicker-autohide
-                               datepicker-format="dd.mm.yyyy"
                                datepicker-button="false"
                                wire:model.live="activityForm.scheduled_period_end"
                         />
@@ -215,12 +213,12 @@
             <div class="flex gap-4 items-end mb-6">
                 <div class="flex-1">
                     <label class="label">Оберіть клінічний запис пацієнта</label>
-                    <select x-model="selectedGround" 
-                            @change="if(selectedGround) { 
+                    <select x-model="selectedGround"
+                            @change="if(selectedGround) {
                                 let parts = selectedGround.split('|');
                                 $wire.addLinkedGround(parts[0], parts[1]);
                                 selectedGround = '';
-                            }" 
+                            }"
                             class="input-select peer w-full">
                         <option value="">-- Оберіть запис --</option>
                         @if(!empty($availableConditions))
