+++ routes/web.php
@@ -23,11 +23,7 @@ use App\Livewire\ContractRequest\ContractRequestEdit;
 use App\Livewire\ContractRequest\ContractRequestIndex;
 use App\Livewire\ContractRequest\ContractRequestShow;
 use App\Livewire\Dashboard;
-use App\Livewire\Declaration\DeclarationCreate;
-use App\Livewire\Declaration\DeclarationEdit;
 use App\Livewire\Declaration\DeclarationIndex;
-use App\Livewire\Declaration\DeclarationView;
-use App\Livewire\DiagnosticReport\DiagnosticReportCreate;
 use App\Livewire\Division\DivisionCreate;
 use App\Livewire\Division\DivisionEdit;
 use App\Livewire\Division\DivisionIndex;
@@ -47,8 +43,6 @@ use App\Livewire\Employee\EmployeeShow;
 use App\Livewire\EmployeeRequest\EmployeeRequestIndex;
 use App\Livewire\EmployeeRole\EmployeeRoleCreate;
 use App\Livewire\EmployeeRole\EmployeeRoleIndex;
-use App\Livewire\Encounter\EncounterCreate;
-use App\Livewire\Encounter\EncounterEdit;
 use App\Livewire\Equipment\EquipmentCreate;
 use App\Livewire\Equipment\EquipmentEdit;
 use App\Livewire\Equipment\EquipmentIndex;
@@ -63,35 +57,13 @@ use App\Livewire\License\LicenseView;
 use App\Livewire\Party\PartyEdit;
 use App\Livewire\Party\PartyVerify;
 use App\Livewire\Party\PartyVerificationIndex;
-use App\Livewire\Person\PersonCreate;
-use App\Livewire\Person\PersonUpdate;
-use App\Livewire\Person\PersonRequestEdit;
-use App\Livewire\Person\PersonIndex;
-use App\Livewire\Person\Records\PatientData;
-use App\Livewire\Person\Records\PatientEpisodes;
-use App\Livewire\Person\Records\PatientSummary;
-use App\Livewire\Person\Records\PatientImmunization;
-use App\Livewire\Person\Records\PatientDiagnoses;
-use App\Livewire\Person\Records\PatientCondition;
-use App\Livewire\Person\Records\PatientDiagnosticReports;
-use App\Livewire\Person\Records\PatientEncounters;
-use App\Livewire\Person\Records\PatientClinicalImpressions;
-use App\Livewire\Person\Records\PatientObservation;
-use App\Livewire\Procedure\ProcedureCreate;
 use App\Models\Declaration;
-use App\Models\DeclarationRequest;
 use App\Models\Division;
 use App\Models\EmployeeRole;
 use App\Models\Equipment;
 use App\Models\HealthcareService;
 use App\Models\LegalEntity;
 use App\Models\License;
-use App\Models\MedicalEvents\Sql\DiagnosticReport;
-use App\Models\MedicalEvents\Sql\Encounter;
-use App\Models\MedicalEvents\Sql\Episode;
-use App\Models\MedicalEvents\Sql\Procedure;
-use App\Models\Person\Person;
-use App\Models\Person\PersonRequest;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -291,61 +263,7 @@ Route::middleware(['auth:ehealth', 'verified'])->group(function () {
                 ->name('declaration.index')
                 ->can('viewAny', Declaration::class);
 
-            Route::prefix('persons')->group(static function () {
-                Route::name('persons.')->group(static function () {
-                    Route::get('/', PersonIndex::class)->can('viewAny', Person::class)->name('index');
-                    Route::get('/create', PersonCreate::class)->can('create', PersonRequest::class)->name('create');
-                    Route::get('/edit/{personRequest}', PersonRequestEdit::class)->can('create', PersonRequest::class)->name('edit');
-                    Route::get('/update/{person}', PersonUpdate::class)->can('create', PersonRequest::class)->name('update');
-
-                    Route::middleware('can:view,' . Person::class)->group(function () {
-                        Route::get('/{personId}/patient-data', PatientData::class)->name('patient-data');
-                        Route::get('/{personId}/summary', PatientSummary::class)->can('view', Person::class)->name('summary');
-                        Route::get('/{personId}/episodes', PatientEpisodes::class)->can('view', Episode::class)->name('episodes');
-                        Route::get('/{personId}/care-plans', \App\Livewire\Person\Records\PersonCarePlans::class)->name('care-plans');
-                        Route::get('/{personId}/observations', PatientObservation::class)->name('observations');
-                        Route::get('/{personId}/immunization', PatientImmunization::class)->name('immunization');
-                        Route::get('/{personId}/condition', PatientCondition::class)->name('condition');
-                        Route::get('/{personId}/diagnoses', PatientDiagnoses::class)->name('diagnoses');
-                        Route::get('/{personId}/diagnostic-reports', PatientDiagnosticReports::class)->name('diagnostic-reports');
-                        Route::get('/{personId}/clinical-impressions', PatientClinicalImpressions::class)->name('clinical-impressions');
-                        Route::get('/{personId}/encounters', PatientEncounters::class)->name('encounters');
-                    });
-                });
-
-                Route::name('declaration.')->group(static function () {
-                    Route::get('/declaration/{declaration}', DeclarationView::class)
-                        ->can('view', 'declaration')
-                        ->name('view')
-                        ->whereNumber('declaration');
-                    Route::get('/{personId}/declaration/create', DeclarationCreate::class)
-                        ->name('create')
-                        ->can('create', DeclarationRequest::class)
-                        ->whereNumber('personId');
-                    Route::get('/{personId}/declaration/{declarationRequest}', DeclarationEdit::class)
-                        ->name('edit')
-                        ->can('update', 'declarationRequest')
-                        ->whereNumber(['personId', 'declarationRequest']);
-                });
-
-                Route::middleware('can:create,' . Encounter::class)->name('encounter.')->group(function () {
-                    Route::get('/{personId}/encounter/create', EncounterCreate::class)->name('create');
-                    Route::get('/{personId}/encounter/{encounterId}', EncounterEdit::class)->name('edit');
-                });
-
-                Route::get('/{personId}/care-plan/create', \App\Livewire\CarePlan\CarePlanCreate::class)
-                    ->name('care-plan.create');
-
-                Route::whereNumber('personId')->group(static function () {
-                    Route::get('{personId}/diagnostic-report/create', DiagnosticReportCreate::class)
-                        ->can('create', DiagnosticReport::class)
-                        ->name('diagnostic-report.create');
-
-                    Route::get('{personId}/procedure/create', ProcedureCreate::class)
-                        ->can('create', Procedure::class)
-                        ->name('procedure.create');
-                });
-            });
+            require __DIR__ . '/persons.php';
         });
 });
